pyproject.toml
[project]
name = "rosehack2026"
version = "0.1.0"
description = "Financial Agent + RAG Pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "openai-agents[litellm]>=0.7.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "fastapi",
    "uvicorn",
    "chromadb",
    "langchain",
    "langchain-community",
    "langchain-google-genai",
    "pypdf",
    "python-multipart",
    "pydantic-settings",
    "uvloop",
    "httptools",
    "presidio-analyzer",
    "presidio-anonymizer",
    "spacy",
    "langgraph",
    "tavily-python",
    "langchain-tavily",
    "langchain-groq",
    "pyjwt",
    "scikit-learn>=1.8.0",
    "pandas>=3.0.0",
    "numpy>=2.4.1",
    "yfinance>=1.1.0",
    "finnhub-python>=2.4.26",
    "cvxpy>=1.7.5",
    "sqlalchemy>=2.0.46",
    "alembic>=1.18.1",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "litellm>=1.0.0",
    "supabase>=2.27.2",
    "wolframclient>=1.4.0",
    "en-core-web-sm",
    "transformers>=4.40.0",
    "torch>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
